requirements.txt
groq
langchain
openai
assemblyai
langchain_community
python-dotenv
langchain-groq
streamlit

sentence-transformers 
scikit-learn